# backend/docs/swagger.yaml
basePath: /
definitions:
...
  models.BookStatusResponse:
    properties:
      reason:
        description: set when rejected
        type: string
      status:
        description: draft, pending, approved, rejected, published
        type: string
    type: object
...
  models.ErrorResponse:
    properties:
      errors:
        additionalProperties:
          type: string
        type: object
      message:
        type: string
    type: object
...
  models.ModerationRequest:
    properties:
      author:
        type: string
      book_id:
        type: string
      description:
        type: string
      is_public:
        type: boolean
      request_id:
        type: string
      title:
        type: string
    type: object
  models.ReviewRequestBody:
    properties:
      reason:
        type: string
      status:
        enum:
        - approved
        - rejected
        type: string
    required:
    - status
    type: object
  models.ReviewRequestResponse:
    properties:
      reason:
        type: string
      status:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  title: Readness API
  version: "1.0"
paths:
...
  /books/{id}/download:
    get:
      description: Только для авторизованных. Видимость как у самой книги
      parameters:
      - description: ID книги
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/pdf
      responses:
        "200":
          description: OK
          schema:
            type: file
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Book or file not found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Скачивание PDF
      tags:
      - Books
  /books/{id}/file:
    put:
      consumes:
      - multipart/form-data
      description: 'multipart/form-data, поле file: PDF до 50 МБ. Права как у редактирования
        книги. Заменяет прежний файл'
      parameters:
      - description: ID книги
        in: path
        name: id
        required: true
        type: string
      - description: PDF-файл
        in: formData
        name: file
        required: true
        type: file
      responses:
        "204":
          description: Файл сохранён
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "413":
          description: Request Entity Too Large
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Загрузка PDF книги
      tags:
      - Books
  /books/{id}/publish:
    post:
      description: Модератор публикует книгу без заявки (закрывает открытую заявку
        как approved)
      parameters:
      - description: ID книги
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: Опубликовано
        "400":
          description: No PDF uploaded
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Прямая публикация
      tags:
      - Book requests
  /books/{id}/publish-request:
    post:
      description: Владелец подаёт заявку на публикацию своей книги; PDF уже должен
        быть загружен
      parameters:
      - description: ID книги
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.BookStatusResponse'
        "400":
          description: No PDF uploaded
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "409":
          description: Already published or request already pending
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Заявка на публикацию
      tags:
      - Book requests
  /books/{id}/status:
    get:
      description: draft — нет заявки, pending/approved/rejected — по последней заявке
        (при rejected есть reason), published — книга в каталоге. Видимость как у
        книги
      parameters:
      - description: ID книги
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.BookStatusResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Статус книги
      tags:
      - Books
  /moderation/requests:
    get:
      description: Открытые заявки, старые первыми (только Moderator+)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ModerationRequest'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Заявки на публикацию
      tags:
      - Book requests
  /moderation/requests/{id}:
    patch:
      consumes:
      - application/json
      description: approved публикует книгу, rejected требует reason (только Moderator+)
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: string
      - description: Решение
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.ReviewRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ReviewRequestResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Pending request not found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Решение по заявке
      tags:
      - Book requests
...
swagger: "2.0"